requirements.txt
python-telegram-bot==20.7
anthropic==0.25.0
apscheduler==3.10.4
pytz==2024.1
